Add test for spl-memo
